requirements.txt
PyYAML==6.0
requests==2.28.0
rich==12.5.1
